requirements.txt
tqdm>=4.64.0
Pillow>=9.0.0
openai>=0.27.0